fix: add watchJSActionSagas in EE file
